refactor recent changes
