+++ WinPort/CMakeLists.txt
@@ -31,6 +31,7 @@ src/Backend/WinPortRGB.cpp
 src/Backend/SudoAskpassImpl.cpp
 src/Backend/FSClipboardBackend.cpp
 src/Backend/ExtClipboardBackend.cpp
+src/Backend/NotifySh.cpp
 src/Backend/TTY/TTYBackend.cpp
 src/Backend/TTY/TTYRevive.cpp
 src/Backend/TTY/TTYInput.cpp
