+++ WinPort/src/Backend/NotifySh.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "WinCompat.h"
+
+SHAREDSYMBOL void Far2l_NotifySh(const char *far2l_path, const char *title, const char *text);
